requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.6
coverage==7.2.7
exceptiongroup==1.1.2
Flask==2.3.2
Flask-JWT-Extended==4.5.2
flask-sqlalchemy==3.0.5
greenlet==2.0.2
gunicorn==21.2.0
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
pytest==7.4.0
python-dotenv==1.0.0
SQLAlchemy==2.0.19
tomli==2.0.1
typing-extensions==4.7.1
Werkzeug==2.3.6
zipp==3.16.2
